# src/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Federation Management Service
  description: |
    # Introduction
    ---
    RESTful APIs that allow an OP to share the edge cloud resources and capabilities securely to other partner OPs over E/WBI.

    ---
    # API Scope

    ---
    APIs defined in this version of the implementation can be categorized into the following areas:
    * __FederationManagement__ - Create and manage directed federation relationship with a partner OP
    * __AvailabilityZoneInfoSynchronization__ - Management of resources of partner OP zones and status updates
    * __ArtefactManagement__ - Upload, remove, retrieve and update application descriptors, charts and packages over E/WBI towards a partner OP
    * __ApplicationOnboardingManagement__ - Register, retrieve, update and remove applications over E/WBI towards a partner OP
    * __ApplicationDeploymentManagement__ - Create, update, retrieve and terminate application instances over E/WBI towards a partner OP

    ---
    # Definitions
    ---
    This section provides definitions of terminologies commonly referred to throughout the API descriptions.

    * __Accepted Zones__ - List of partner OP zones, which the originating OP has confirmed to use for its edge applications.
    * __Application Provider__ - An application developer, onboarding his/her edge application on a partner operator platform (MEC).
    * __Artefact__ - Descriptor, charts or any other package associated with the application.
...

    ---
...

...

...

...

...

...


...
servers:
...
security:
- oAuth2ClientCredentials:
  - fed-mgmt
paths:
...
  /{federationContextId}/zones:
    post:
      tags:
      - AvailabilityZoneInfoSynchronization
      summary: Originating OP informs partner OP that it is willing to access the
        specified zones and partner OP shall reserve compute and network resources
        for these zones.
      operationId: zone_subscribe
      parameters:
      - name: federationContextId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/FederationContextId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZoneRegistrationRequestData'
        required: true
      responses:
        "200":
          description: Zone registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZoneRegistrationResponseData'
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "422":
          description: Unprocessable Entity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "503":
          description: Service Unavailable
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        "520":
          description: Web Server Returned an Unknown Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        default:
          description: Generic Error
      callbacks:
        onZoneResourceUpdateEvent:
          '{$request.body#/availZoneNotifLink}':
            post:
              requestBody:
                description: Notification about resource availability.
                content:
                  application/json:
                    schema:
                      required:
                      - federationContextId
                      - zoneId
                      - zoneResUpdInfo
                      type: object
                      properties:
                        federationContextId:
                          $ref: '#/components/schemas/FederationIdentifier'
                        zoneId:
                          $ref: '#/components/schemas/ZoneIdentifier'
                        zoneResUpdInfo:
                          type: array
                          items:
                            minProperties: 1
                            type: object
                            properties:
                              availableCompResources:
                                minItems: 1
                                type: array
                                description: Resources exclusively reserved for the
                                  originator OP.
                                items:
                                  $ref: '#/components/schemas/ComputeResourceInfo'
                              availableNetResources:
                                minProperties: 1
                                type: object
                                properties:
                                  egressBandWidth:
                                    type: integer
                                    description: Max dl throughput that this edge
                                      can offer. It is defined in Mbps.
                                    format: int32
                                  dedicatedNIC:
                                    type: integer
                                    format: int32
                                  supportSriov:
                                    type: boolean
                                    description: If this zone support SRIOV networks
                                      or not
                                  supportDPDK:
                                    type: boolean
                                    description: If this zone supports DPDK based
                                      networking
              responses:
                "200":
                  description: Zone info notification acknowledged
                "400":
                  description: Bad request
                  content:
                    application/problem+json:
                      schema:
                        $ref: '#/components/schemas/ProblemDetails'
                "401":
                  description: Unauthorized
                  content:
                    application/problem+json:
                      schema:
                        $ref: '#/components/schemas/ProblemDetails'
                "404":
                  description: Not Found
                  content:
                    application/problem+json:
                      schema:
                        $ref: '#/components/schemas/ProblemDetails'
                "409":
                  description: Conflict
                  content:
                    application/problem+json:
                      schema:
                        $ref: '#/components/schemas/ProblemDetails'
                "422":
                  description: Unprocessable Entity
                  content:
                    application/problem+json:
                      schema:
                        $ref: '#/components/schemas/ProblemDetails'
                "500":
                  description: Internal Server Error
                  content:
                    application/problem+json:
                      schema:
                        $ref: '#/components/schemas/ProblemDetails'
                "503":
                  description: Service Unavailable
                  content:
                    application/problem+json:
                      schema:
                        $ref: '#/components/schemas/ProblemDetails'
                "520":
                  description: Web Server Returned an Unknown Error
                  content:
                    application/problem+json:
                      schema:
                        $ref: '#/components/schemas/ProblemDetails'
                default:
                  description: Generic Error
      x-openapi-router-controller: controllers.availability_zone_info_synchronization_controller
...
components:
  schemas:
...
    FederationContextId:
      pattern: "^[A-Za-z0-9][A-Za-z0-9-]*$"
      type: string
      description: This identifier shall be provided by the partner OP on successful
        verification and validation of the federation create request and is used by
        partner op to identify this newly created federation context. Originating
        OP shall provide this identifier in any subsequent request towards the partner
        op.
      readOnly: true
    FederationIdentifier:
      pattern: "^[A-Za-z0-9][A-Za-z0-9-]*$"
      type: string
      description: Globally unique identifier allocated to an operator platform. This
        is valid and used only in context of MEC federation interface.
...
    Uri:
      type: string
    Vcpu:
      pattern: "^\\d+((\\.\\d{1,3})|(m))?$"
      type: string
      description: "Number of vcpus in whole, decimal up to millivcpu, or millivcpu\
        \ format."
      example: "{\"whole\":{\"value\":2},\"decimal\":{\"value\":0.5},\"millivcpu\"\
        :{\"value\":\"500m\"}}"
...
    ZoneIdentifier:
      pattern: "^[A-Za-z0-9][A-Za-z0-9-]*$"
      type: string
      description: Human readable name of the zone.
...
    ComputeResourceInfo:
      required:
      - cpuArchType
      - memory
      - numCPU
      type: object
      properties:
        cpuArchType:
          type: string
          description: "CPU Instruction Set Architecture (ISA) E.g., Intel, Arm etc."
          enum:
          - ISA_X86_64
          - ISA_ARM_64
        numCPU:
          $ref: '#/components/schemas/Vcpu'
        memory:
          type: integer
          description: Amount of RAM in Mbytes
          format: int64
        diskStorage:
          type: integer
          description: Amount of disk storage in Gbytes for a given ISA type
          format: int32
        gpu:
          type: array
          items:
            $ref: '#/components/schemas/GpuInfo'
        vpu:
          type: integer
          description: Number of Intel VPUs available for a given ISA type
        fpga:
          type: integer
          description: Number of FPGAs available for a given ISA type
        hugepages:
          type: array
          items:
            $ref: '#/components/schemas/HugePage'
        cpuExclusivity:
          type: boolean
          description: Support for exclusive CPUs
      example:
        hugepages:
        - number: 7
          pageSize: 2MB
        - number: 7
          pageSize: 2MB
        memory: 0
        fpga: 2
        cpuArchType: ISA_X86_64
        diskStorage: 6
        numCPU: "{\"whole\":{\"value\":2},\"decimal\":{\"value\":0.5},\"millivcpu\"\
          :{\"value\":\"500m\"}}"
        vpu: 5
        cpuExclusivity: true
        gpu:
        - gpuMemory: 1
          gpuVendorType: Nvidia
          gpuModeName: gpuModeName
          numGPU: 5
        - gpuMemory: 1
          gpuVendorType: Nvidia
          gpuModeName: gpuModeName
          numGPU: 5
...
    GpuInfo:
      required:
      - gpuMemory
      - gpuModeName
      - gpuVendorType
      - numGPU
      type: object
      properties:
        gpuVendorType:
          type: string
          description: "GPU vendor name e.g. NVIDIA, AMD etc."
          example: Nvidia
          enum:
          - GPU_PROVIDER_NVIDIA
          - GPU_PROVIDER_AMD
        gpuModeName:
          type: string
          description: "Model name corresponding to vendorType may include info e.g.\
            \ for NVIDIA, model name could be “Tesla M60”, “Tesla V100” etc."
        gpuMemory:
          type: integer
          description: GPU memory in Mbytes
        numGPU:
          type: integer
          description: Number of GPUs
      example:
        gpuMemory: 1
        gpuVendorType: Nvidia
        gpuModeName: gpuModeName
        numGPU: 5
...
    InvalidParam:
      required:
      - param
      type: object
      properties:
        param:
          type: string
        reason:
          type: string
...
    ProblemDetails:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        cause:
          type: string
        invalidParams:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/InvalidParam'
...
    ZoneRegistrationRequestData:
      required:
      - acceptedAvailabilityZones
      - availZoneNotifLink
      type: object
      properties:
        acceptedAvailabilityZones:
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/ZoneIdentifier'
        availZoneNotifLink:
          $ref: '#/components/schemas/Uri'
...
  securitySchemes:
    oAuth2ClientCredentials:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: http://127.0.0.1:8080/realms/federation/protocol/openid-connect/token
          scopes:
            fed-mgmt: Access to the federation APIs
      x-tokenInfoFunc: controllers.authorization_controller.check_oAuth2ClientCredentials
      x-scopeValidateFunc: controllers.authorization_controller.validate_scope_oAuth2ClientCredentials